+++ packages/iov-bcp-types/types/transactions.d.ts
@@ -133,7 +133,7 @@ export interface SendTransaction extends UnsignedTransaction {
 }
 export interface SwapOfferTransaction extends UnsignedTransaction {
     readonly kind: "bcp/swap_offer";
-    readonly amount: ReadonlyArray<Amount>;
+    readonly amounts: ReadonlyArray<Amount>;
     readonly recipient: Address;
     /** absolute block height at which the offer times out */
     readonly timeout: number;
@@ -141,7 +141,7 @@ export interface SwapOfferTransaction extends UnsignedTransaction {
 }
 export interface SwapCounterTransaction extends UnsignedTransaction {
     readonly kind: "bcp/swap_counter";
-    readonly amount: ReadonlyArray<Amount>;
+    readonly amounts: ReadonlyArray<Amount>;
     readonly recipient: Address;
     /** absolute block height at which the counter offer times out */
     readonly timeout: number;